refactor: replace BitcoinUrlHandler with LitecoinURIHandler
This is a Litecoin wallet, and BitcoinUrlHandler's litecoin: URI/QR handling
logic was a byte-for-byte duplicate of the already-present but unused
LitecoinURIHandler - a leftover from an incomplete Bitcoin -> Litecoin naming
pass. Consolidate on the Litecoin-named class, part of the long-term goal of
referencing Litecoin (not Bitcoin) throughout the codebase.

- LitecoinURIHandler now carries the full logic: processRequest,
  tryPaymentRequest, tryLitecoinURL (including the deep-link-to-Send-screen
  fix from the previous commit), plus its own pre-existing
  isValidLitecoinURI/isValidLitecoinUrl/getRequestFromString.
- BreadActivity, ScanQRActivity, and BRActivity now call
  LitecoinURIHandler.processRequest / isValidLitecoinURI instead of
  BitcoinUrlHandler.processRequest / isBitcoinUrl.
- BitcoinUrlHandler is reduced to just its three native method declarations
  (parsePaymentRequest, parsePaymentACK, getCertificatesFromPaymentRequest).
  Those can't move: the native library resolves them by their exact JNI
  symbol name (Java_com_brainwallet_tools_security_BitcoinUrlHandler_...,
  see app/src/main/jni/transition/core.c), which lives in the `core` git
  submodule - renaming the Java class without a matching native-side change
  would break payment-protocol requests at runtime with an
  UnsatisfiedLinkError. PaymentProtocolTask and X509CertificateValidator
  keep calling these directly, unchanged. Left a comment on the class
  explaining why it's still there so it doesn't look like accidental
  leftover dead code.

Verified: compileBrainwalletDebugKotlin, compileBrainwalletDebugJavaWithJavac,
detekt, testBrainwalletDebugUnitTest, and assembleBrainwalletDebug all pass
clean.
