+++ app/src/main/java/com/brainwallet/tools/security/BitcoinUrlHandler.java
@@ -1,169 +1,19 @@
 package com.brainwallet.tools.security;
-import androidx.fragment.app.FragmentActivity;
-import com.brainwallet.tools.animation.BRDialog;
-import com.brainwallet.tools.threads.PaymentProtocolTask;
-import com.brainwallet.R;
-import com.brainwallet.navigation.LegacyNavigation;
-import com.brainwallet.navigation.Route;
-import com.brainwallet.presenter.customviews.BRDialogView;
+
 import com.brainwallet.presenter.entities.PaymentRequestWrapper;
-import com.brainwallet.presenter.entities.RequestObject;
-import com.brainwallet.util.EventBus;
-import com.brainwallet.wallet.BRWalletManager;
 
-import java.io.UnsupportedEncodingException;
-import java.math.BigDecimal;
-import java.net.URI;
-import java.net.URLDecoder;
-
-import timber.log.Timber;
-
+/**
+ * BIP70 payment-protocol native bindings only.
+ *
+ * The higher-level litecoin: URI/QR handling this class used to own has moved to
+ * {@link LitecoinURIHandler} - use that instead for anything new. This class stays
+ * around solely to host these three native methods: the native library resolves them
+ * by their fully-qualified JNI symbol name
+ * (`Java_com_brainwallet_tools_security_BitcoinUrlHandler_...`, see
+ * app/src/main/jni/transition/core.c, which lives in the `core` git submodule), so
+ * renaming or moving them requires a matching change there first.
+ */
 public class BitcoinUrlHandler {
-    private static final Object lockObject = new Object();
-
-    public static synchronized boolean processRequest(FragmentActivity app, String url) {
-        if (url == null) {
-            Timber.d("timber: processRequest: url is null");
-            return false;
-        }
-
-        RequestObject requestObject = getRequestFromString(url);
-        if (BRWalletManager.getInstance().confirmSweep(app, url)) {
-            return true;
-        }
-        if (requestObject == null) {
-            if (app != null) {
-                BRDialog.showCustomDialog(app, app.getString(R.string.JailbreakWarnings_title),
-                        app.getString(R.string.Send_invalidAddressTitle), app.getString(R.string.Button_ok), null, new BRDialogView.BROnClickListener() {
-                            @Override
-                            public void onClick(BRDialogView brDialogView) {
-                                brDialogView.dismissWithAnimation();
-                            }
-                        }, null, null, 0);
-            }
-            return false;
-        }
-        if (requestObject.r != null) {
-            return tryPaymentRequest(requestObject);
-        } else if (requestObject.address != null) {
-            return tryLitecoinURL(url, app);
-        } else {
-            if (app != null) {
-                BRDialog.showCustomDialog(app, app.getString(R.string.JailbreakWarnings_title),
-                        app.getString(R.string.Send_remoteRequestError), app.getString(R.string.Button_ok), null, new BRDialogView.BROnClickListener() {
-                            @Override
-                            public void onClick(BRDialogView brDialogView) {
-                                brDialogView.dismissWithAnimation();
-                            }
-                        }, null, null, 0);
-            }
-            return false;
-        }
-    }
-
-    public static boolean isBitcoinUrl(String url) {
-        RequestObject requestObject = getRequestFromString(url);
-        // return true if the request is valid url and has param: r or param: address
-        // return true if it is a valid bitcoinPrivKey
-        return (requestObject != null && (requestObject.r != null || requestObject.address != null)
-                || BRWalletManager.getInstance().isValidBitcoinBIP38Key(url)
-                || BRWalletManager.getInstance().isValidBitcoinPrivateKey(url));
-    }
-
-
-    public static RequestObject getRequestFromString(String str) {
-        if (str == null || str.isEmpty()) return null;
-        RequestObject obj = new RequestObject();
-
-        String tmp = str.trim().replaceAll("\n", "").replaceAll(" ", "%20");
-
-        if (!tmp.startsWith("litecoin://")) {
-            if (!tmp.startsWith("litecoin:"))
-                tmp = "litecoin://".concat(tmp);
-            else
-                tmp = tmp.replace("litecoin:", "litecoin://");
-        }
-        URI uri;
-        try {
-            uri = URI.create(tmp);
-        } catch (IllegalArgumentException ex) {
-            Timber.e(ex, "getRequestFromString: ");
-            return null;
-        }
-
-        String host = uri.getHost();
-        if (host != null) {
-            String addrs = host.trim();
-            if (BRWalletManager.getInstance().validateAddress(addrs)) {
-                obj.address = addrs;
-            }
-        }
-        String query = uri.getQuery();
-        if (query == null) return obj;
-        String[] params = query.split("&");
-        for (String s : params) {
-            String[] keyValue = s.split("=", 2);
-            if (keyValue.length != 2)
-                continue;
-            if (keyValue[0].trim().equals("amount")) {
-                try {
-                    BigDecimal bigDecimal = new BigDecimal(keyValue[1].trim());
-                    obj.amount = bigDecimal.multiply(new BigDecimal("100000000")).toString();
-                } catch (NumberFormatException e) {
-                    Timber.e(e);
-                }
-            } else if (keyValue[0].trim().equals("label")) {
-                obj.label = keyValue[1].trim();
-            } else if (keyValue[0].trim().equals("message")) {
-                obj.message = keyValue[1].trim();
-            } else if (keyValue[0].trim().startsWith("req")) {
-                obj.req = keyValue[1].trim();
-            } else if (keyValue[0].trim().startsWith("r")) {
-                obj.r = keyValue[1].trim();
-            }
-        }
-        return obj;
-    }
-
-    private static boolean tryPaymentRequest(RequestObject requestObject) {
-        String theURL;
-        String url = requestObject.r;
-        synchronized (lockObject) {
-            try {
-                theURL = URLDecoder.decode(url, "UTF-8");
-            } catch (UnsupportedEncodingException e) {
-                Timber.e(e);
-                return false;
-            }
-            new PaymentProtocolTask().execute(theURL, requestObject.label);
-        }
-        return true;
-    }
-
-    private static boolean tryLitecoinURL(final String url, final FragmentActivity app) {
-        RequestObject requestObject = getRequestFromString(url);
-        if (requestObject == null || requestObject.address == null || requestObject.address.isEmpty())
-            return false;
-
-        String amount = requestObject.amount;
-
-        if (amount == null || amount.isEmpty() || new BigDecimal(amount).doubleValue() == 0) {
-            app.runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    // Keep posting this in case the Send screen is already open and
-                    // collecting (e.g. address scanned via the in-app camera flow).
-                    EventBus.INSTANCE.postQRCodeScanned(requestObject.address);
-                    // Deep link straight to the Send screen with the address pasted
-                    // in, so scanning this QR code from another app (camera, another
-                    // wallet, etc.) reliably lands there instead of wherever the app
-                    // happened to be, or nowhere at all on a cold start.
-                    LegacyNavigation.openComposeScreen(app, new Route.Send(requestObject.address));
-                }
-            });
-        }
-        return true;
-    }
 
     public static native PaymentRequestWrapper parsePaymentRequest(byte[] req);
