+++ app/src/main/java/com/brainwallet/tools/security/LitecoinURIHandler.java
@@ -1,22 +1,74 @@
 package com.brainwallet.tools.security;
 
+import androidx.fragment.app.FragmentActivity;
+
 import com.brainwallet.R;
+import com.brainwallet.navigation.LegacyNavigation;
+import com.brainwallet.navigation.Route;
 import com.brainwallet.presenter.customviews.BRDialogView;
-import com.brainwallet.presenter.entities.PaymentRequestWrapper;
 import com.brainwallet.presenter.entities.RequestObject;
 import com.brainwallet.tools.animation.BRDialog;
 import com.brainwallet.tools.threads.PaymentProtocolTask;
 import com.brainwallet.util.EventBus;
 import com.brainwallet.wallet.BRWalletManager;
+
 import java.io.UnsupportedEncodingException;
 import java.math.BigDecimal;
 import java.net.URI;
 import java.net.URLDecoder;
+
 import timber.log.Timber;
 
-import androidx.fragment.app.FragmentActivity;
-
+/**
+ * Handles incoming litecoin: URIs (deep links, QR scans) - address-only requests,
+ * BIP21-style amount/label/message params, and BIP70 payment-protocol requests (r=).
+ *
+ * This is the Litecoin-named entry point for that logic; use it instead of
+ * {@link BitcoinUrlHandler}, which now exists only to hold the payment-protocol
+ * native method declarations (see that class for why).
+ */
 public class LitecoinURIHandler {
+    private static final Object lockObject = new Object();
+
+    public static synchronized boolean processRequest(FragmentActivity app, String url) {
+        if (url == null) {
+            Timber.d("timber: processRequest: url is null");
+            return false;
+        }
+
+        RequestObject requestObject = getRequestFromString(url);
+        if (BRWalletManager.getInstance().confirmSweep(app, url)) {
+            return true;
+        }
+        if (requestObject == null) {
+            if (app != null) {
+                BRDialog.showCustomDialog(app, app.getString(R.string.JailbreakWarnings_title),
+                        app.getString(R.string.Send_invalidAddressTitle), app.getString(R.string.Button_ok), null, new BRDialogView.BROnClickListener() {
+                            @Override
+                            public void onClick(BRDialogView brDialogView) {
+                                brDialogView.dismissWithAnimation();
+                            }
+                        }, null, null, 0);
+            }
+            return false;
+        }
+        if (requestObject.r != null) {
+            return tryPaymentRequest(requestObject);
+        } else if (requestObject.address != null) {
+            return tryLitecoinURL(url, app);
+        } else {
+            if (app != null) {
+                BRDialog.showCustomDialog(app, app.getString(R.string.JailbreakWarnings_title),
+                        app.getString(R.string.Send_remoteRequestError), app.getString(R.string.Button_ok), null, new BRDialogView.BROnClickListener() {
+                            @Override
+                            public void onClick(BRDialogView brDialogView) {
+                                brDialogView.dismissWithAnimation();
+                            }
+                        }, null, null, 0);
+            }
+            return false;
+        }
+    }
 
     public static RequestObject getRequestFromString(String str) {
         if (str == null || str.isEmpty()) return null;
@@ -72,13 +124,15 @@ public class LitecoinURIHandler {
         return obj;
     }
 
+    /**
+     * Returns true if the request is a valid URI with an r= or address param, or a
+     * valid Litecoin BIP38/private key.
+     */
     public static boolean isValidLitecoinURI(String url) {
         RequestObject requestObject = getRequestFromString(url);
-        // return true if the request is valid url and has param: r or param: address
-        // return true if it is a valid bitcoinPrivKey
         return (requestObject != null && (requestObject.r != null || requestObject.address != null)
-            || BRWalletManager.getInstance().isValidBitcoinBIP38Key(url)
-            || BRWalletManager.getInstance().isValidBitcoinPrivateKey(url));
+                || BRWalletManager.getInstance().isValidBitcoinBIP38Key(url)
+                || BRWalletManager.getInstance().isValidBitcoinPrivateKey(url));
     }
 
     public static boolean isValidLitecoinUrl(String url) {
@@ -86,6 +140,21 @@ public class LitecoinURIHandler {
             return BRWalletManager.getInstance().validateAddress(url);
         }
         return false;
+    }
+
+    private static boolean tryPaymentRequest(RequestObject requestObject) {
+        String theURL;
+        String url = requestObject.r;
+        synchronized (lockObject) {
+            try {
+                theURL = URLDecoder.decode(url, "UTF-8");
+            } catch (UnsupportedEncodingException e) {
+                Timber.e(e);
+                return false;
+            }
+            new PaymentProtocolTask().execute(theURL, requestObject.label);
+        }
+        return true;
     }
 
     private static boolean tryLitecoinURL(final String url, final FragmentActivity app) {
@@ -99,12 +168,17 @@ public class LitecoinURIHandler {
             app.runOnUiThread(new Runnable() {
                 @Override
                 public void run() {
+                    // Keep posting this in case the Send screen is already open and
+                    // collecting (e.g. address scanned via the in-app camera flow).
                     EventBus.INSTANCE.postQRCodeScanned(requestObject.address);
+                    // Deep link straight to the Send screen with the address pasted
+                    // in, so scanning this QR code from another app (camera, another
+                    // wallet, etc.) reliably lands there instead of wherever the app
+                    // happened to be, or nowhere at all on a cold start.
+                    LegacyNavigation.openComposeScreen(app, new Route.Send(requestObject.address));
                 }
             });
         }
         return true;
     }
-
-
 }
